### 2018-02-08 #574 @whedon
[PRE REVIEW]: Nostril: A nonsense string evaluator written in Python
**Submitting author:** @mhucka (<a href="http://orcid.org/0000-0001-9105-5960">Michael Hucka</a>)
**Repository:** <a href="https://github.com/casics/nostril" target ="_blank">https://github.com/casics/nostril</a>
**Version:** v1.0.0
