### 2018-02-09 #574 @mhucka
Thanks for all that. Regarding reviewers, I don't know these people but based purely on the list of topic areas in that fantastic spreadsheet, the following appear to be most closely related:

* desilinguist
* ahurriyetoglu
* jkahn
* @kellieotto
* linuxscout
* xirdneh
