### 2018-02-27 #574 @jakevdp
Hi @mhucka, thanks for the submission! Looks really interesting... I'm going to work on finding a reviewer or two and start this process along :smile:
