### 2018-02-27 #596 @whedon
[REVIEW]: Nostril: A nonsense string evaluator written in Python
**Submitting author:** @mhucka (<a href="http://orcid.org/0000-0001-9105-5960">Michael Hucka</a>)
**Repository:** <a href="https://github.com/casics/nostril" target ="_blank">https://github.com/casics/nostril</a>
**Version:** v1.0.0
**Archive:** <a href="http://dx.doi.org/10.22002/D1.935" target="_blank">10.22002/D1.935</a>
