### 2018-02-27 #596 @jakevdp
@desilinguist thanks for agreeing to review! The review will take place here in this issue – all relevant info is above. Please feel free to reach out to me if you have questions about the process!
