### 2018-03-17 39e1c48
Fix import of version info so callers can use "nostril.__version__"
