### 2018-03-23 e9cf463
Add NSF and Caltech logos
