### 2018-03-26 #596 @desilinguist
This is a package that uses information retrieval methods to solve a very niche but interesting problem - figuring out whether a given string is completely nonsense or whether it is likely to be "meaningful text" but specifically in the context of identifier names used in source code. It's a very interesting approach and I quite like it.

In general, the submission is pretty much good to go except for a few things. Here are my suggestions:

1. The installation could be improved. It would be nice to have an actual package hosted on PyPI or on anaconda.org instead of installing from the git repository. It would also be nice to explicitly list the dependencies in the repository README.

2. Although the main readme does state what the primary use case is ("... to decide whether short strings returned by source code mining methods are likely to be program identifiers ... or random characters or other non-identifier strings"), I think this can be improved a little bit. It would be nice tp provide a bit of explanation as to why this is something that someone would need to do. Perhaps have some pointers to the source code mining literature and a brief summary?

3. There is no API-level documentation per se (it could easily be added via readthedocs.org). The documentation link from the main README is broken. It would also be nice to take the very-nicely-written documentation from under `docs/explanations` and move it to readthedocs since right now it's not really easily discoverable.

4. It would be nice to use a real test library instead of having custom functions. Did the authors try other libraries like `nose`, `nose2`, or `unittest`? If it turns out that the custom test functions ( `test_lableled()` and `test_unlabeled()`) are indeed necessary, perhaps they should be moved to a `utils.py` under the `tests` folder since it's odd to have them be part of the main source code.

5. The references in the paper don't seem to have DOIs.

6. The statement of need in the paper can be improved in the same way as I suggested above.

All in all, this is a really nice package and I think it should be accepted to the journal pending some of the changes I suggested above. Thanks for the great work @mhucka!
