### 2018-03-27 #596 @desilinguist
@jakevdp @arfon I am assuming you got notified of this.
