### 2018-03-27 #596 @arfon
> @jakevdp @arfon I am assuming you got notified of this.

👍yep, thanks @desilinguist!
