### 2018-04-16 #596 @arfon
@mhucka - please take another look at this when you get a chance.
