### 2018-04-26 58b08cf
Move main code to new __main__.py
This allows users to run Nostril using "python3 -m nostril".
