### 2018-04-26 #596 @mhucka
Update: I've now fixed one issue, which is that I used an antipattern in `setup.py` that caused dependencies not to be installed automatically. I should be able to respond to the remaining issues this week.
