### 2018-04-27 847e459
Add more information about dependencies
