### 2018-04-27 f83f34f
Using -V instead of -v for version seems more conventional
Looking around at a number of other programs, it seems more
conventional to use -V for the short form of --version, instead of the
lower-case -v that I used before.
