### 2018-04-27 1aa9e6c
Update references and add DOIs where possible
Also switched to using Paperpile's BibTeX output.
