+++ docs/papers/joss/paper.bib
@@ -1,31 +1,39 @@
-@inproceedings{deb2000fast,
-  author = {Deb, Kalyanmoy  and Agrawal, Samir  and Pratap, Amrit  and Meyarivan, Tanaka},
-  booktitle = {International Conference on Parallel Problem Solving From Nature},
-  title = {A fast elitist non-dominated sorting genetic algorithm for multi-objective optimization: {NSGA}-{I}{I}},
-  pages = {849-858},
-  year = {2000},
-  url = {},
-  tags = {casics}
+@INPROCEEDINGS{deb2000fast,
+  title     = "A Fast Elitist Non-dominated Sorting Genetic Algorithm for
+               Multi-objective Optimization: {NSGA-II}",
+  booktitle = "Parallel Problem Solving from Nature {PPSN} {VI}",
+  author    = "Deb, Kalyanmoy and Agrawal, Samir and Pratap, Amrit and
+               Meyarivan, T",
+  publisher = "Springer",
+  pages     = "849--858",
+  year      =  2000,
+  url       = "http://dx.doi.org/10.1007/3-540-45356-3_83",
+  doi       = "10.1007/3-540-45356-3_83"
 }
 
-@book{manning2009introduction,
-  author = {Manning, Christopher D.  and Raghavan, Prabhakar  and Sch\"{u}tze, Hinrich},
-  title = {Introduction to Information Retrieval},
-  publisher = {Cambridge University Press},
-  year = {2009},
-  url = {},
-  tags = {methods/algorithms;text classification;casics},
-  googlebooks_key = {t1PoSh4uwVcC},
-  edition = {Online edition}
+@BOOK{manning2009introduction,
+  title     = "Introduction to Information Retrieval",
+  author    = "Manning, Christopher D and Raghavan, Prabhakar and Sch{\"u}tze,
+               Hinrich",
+  publisher = "Cambridge University Press",
+  edition   = "[Online edition]",
+  year      =  2009,
+  url       = "https://market.android.com/details?id=book-t1PoSh4uwVcC",
+  language  = "en",
+  isbn      = "9781139472104",
+  doi       = "10.1017/CBO9780511809071",
+  url       = "https://doi.org/10.1017/CBO9780511809071"
 }
 
-@inproceedings{binkley2013dataset,
-  author = {Binkley, David  and Lawrie, Dawn  and Pollock, Lori  and Hill, Emily  and Vijay-Shanker, K.},
-  booktitle = {Proceedings of the 10th Working Conference on Mining Software Repositories},
-  title = {A dataset for evaluating identifier splitters},
-  pages = {401-404},
-  year = {2013},
-  url = {},
-  tags = {code-mining-casbium;abstract-done}
+@INPROCEEDINGS{binkley2013dataset,
+  title     = "A Dataset for Evaluating Identifier Splitters",
+  booktitle = "Proceedings of the 10th Working Conference on Mining Software
+               Repositories",
+  author    = "Binkley, David and Lawrie, Dawn and Pollock, Lori and Hill,
+               Emily and Vijay-Shanker, K",
+  publisher = "IEEE Press",
+  pages     = "401--404",
+  year      =  2013,
+  url       = "http://dl.acm.org/citation.cfm?id=2487085.2487158",
+  isbn      = "9781467329361",
 }
-
