### 2018-04-27 8afe814
Add more motivations and more references
