+++ docs/papers/joss/paper.bib
@@ -37,3 +37,33 @@
   url       = "http://dl.acm.org/citation.cfm?id=2487085.2487158",
   isbn      = "9781467329361",
 }
+
+@INPROCEEDINGS{Dit2011-fz,
+  title     = "Can Better Identifier Splitting Techniques Help Feature
+               Location?",
+  booktitle = "2011 {IEEE} 19th International Conference on Program
+               Comprehension",
+  author    = "Dit, B and Guerrouj, L and Poshyvanyk, D and Antoniol, G",
+  pages     = "11--20",
+  month     =  jun,
+  year      =  2011,
+  url       = "http://dx.doi.org/10.1109/ICPC.2011.47",
+  issn      = "1092-8138",
+  doi       = "10.1109/ICPC.2011.47"
+}
+
+@ARTICLE{Linstead2009-ky,
+  title   = "Sourcerer: mining and searching internet-scale software
+             repositories",
+  author  = "Linstead, Erik and Bajracharya, Sushil and Ngo, Trung and Rigor,
+             Paul and Lopes, Cristina and Baldi, Pierre",
+  journal = "Data mining and knowledge discovery",
+  volume  =  18,
+  number  =  2,
+  pages   = "300--336",
+  month   =  apr,
+  year    =  2009,
+  url     = "https://doi.org/10.1007/s10618-008-0118-x",
+  issn    = "1384-5810, 1573-756X",
+  doi     = "10.1007/s10618-008-0118-x"
+}
