### 2018-04-27 8e63c94
Version bump to 1.1
