### 2018-04-27 a638996
Add a News section & bump version to 1.1.0
