### 2018-04-27 985de6a
Fix broken link to explanations document
