### 2018-04-27 8dc449c
Add another link to the longer explanation to make it easier to find
