### 2018-04-27 c0b8b85
Add more information to motivate the need for Nostril
