### 2018-04-27 #596 @mhucka
Thank you for the review and suggestions.  Here are some replies, finally:

> The installation could be improved. It would be nice to have an actual package hosted on PyPI or on anaconda.org instead of installing from the git repository. It would also be nice to explicitly list the dependencies in the repository README.

I have now fixed an antipattern in `setup.py` that caused dependencies not to be properly installed by Python's setuptools, which is undoubtedly what led to the comment about the need to list the dependencies.  The dependencies should now be installed automatically by `setup.py`, and the entire processs should go more smoothly now.  I've updated the installation instructions and also included a mention of the dependencies just for completeness' sake.  Having a package hosted on PyPI is a future goal.

> Although the main readme does state what the primary use case is ("... to decide whether short strings returned by source code mining methods are likely to be program identifiers ... or random characters or other non-identifier strings"), I think this can be improved a little bit. It would be nice tp provide a bit of explanation as to why this is something that someone would need to do. Perhaps have some pointers to the source code mining literature and a brief summary?

The revised readme and the paper now include more motivations on this point, which was admittedly not sufficiently addressed in the first version of the paper.

> There is no API-level documentation per se (it could easily be added via readthedocs.org). The documentation link from the main README is broken. It would also be nice to take the very-nicely-written technical explanation of how the package works from under docs/explanations and move it to readthedocs since right now it's not really easily discoverable.

I've fixed the link to the documentation file in the README.  Putting docs on readthedocs.org is a future goal.

> It would be nice to use a real test library instead of having custom functions. Did the authors try other libraries like nose, nose2, or unittest? If it turns out that the custom test functions ( test_lableled() and test_unlabeled()) are indeed necessary, perhaps they should be moved to a utils.py under the tests folder since it's odd to have them be part of the main source code.

Indeed, it would be nice, but due to limited time I cannot undertake this now, and in any case it is not a publication requirement for JOSS.  However, like some of the other suggestions above, I've added this to the issues in GitHub as a future goal.

> The references in the paper don't seem to have DOIs.

There are now DOIs where possible.  (Not all publications have DOIs.)

> The statement of need in the paper can be improved in the same way as I suggested above.

This is now done.
