### 2018-04-27 #596 @mhucka
Hmm. I don't know why Whedon isn't responding. It worked for one of my other JOSS papers ... @arfon do you have any thoughts about the command above didn't regenerate the pdf?  ([Here](https://github.com/openjournals/joss-reviews/issues/653#issuecomment-377441772]) is an example of where it worked in the context of another review.)
