### 2018-04-29 eef9a61
Don't require specific package versions; use '>=' instead
