+++ docs/explanations/README.md
@@ -1,28 +1,33 @@
 Nostril: Nonsense String Evaluator
 ==================================
 
-This Python 3 module is designed to detect whether a given medium-length string of characters is likely to be random gibberish or something meaningful.  It makes a binary distinction: either a given input is (probably) nonsense, or it is (probably) not.  The main use case is to decide whether strings returned by source code mining methods are likely to be (e.g.) program identifiers, or random characters or other non-identifier strings.  The function generated by this system takes a text string and returns `True` if it is probably nonsense, `False` otherwise (if it likely to be meaningful).  _Nostril_ is reasonably fast: once the module is loaded, on a 4 Ghz Apple OS X 10.12 computer, calling the evaluation function returns a result in 20-30 microseconds on average.
+Nostril (_Nonsense String Evaluator_) is a Python 3 package is designed to detect whether a given medium-length string of characters is likely to be random gibberish or something meaningful.  It makes a binary distinction: either a given input is (probably) nonsense, or it is (probably) not.  The main use case is to decide whether strings returned by source code mining methods are likely to be (e.g.) program identifiers, or random characters or other non-identifier strings.
+
+Table of contents
+-----------------
+
+* [Usage](#usage)
+* [Known limitations](#known-limitations)
+* [Operating principles](#operating-principles)
+* [Tuning, training and testing](#tuning-training-and-testing)
+* [Finding parameter values](#finding-parameter-values)
+* [Other approaches tried](#other-approaches-tried)
 
 Usage
 -----
 
-The basic usage is very simple.  _Nostril_ is a Python module that (among other things) provides a function named `is_nonsense()`.  This function takes a text string as an argument and returns a Boolean value as a result.  Here is an example:
+The basic usage is very simple.  _Nostril_ is a Python package that (among other things) provides a function named `nonsense()`.  This function takes a text string as an argument and returns a Boolean value as a result.  Here is an example:
 
-    from nostril import is_nonsense
-    result = is_nonsense('yoursinglestringhere')
+    from nostril import nonsense
+    for s in ['bunchofwords', 'xywinlist', 'faiwtlwexu', 'asfgtqwafazfy']:
+        if nonsense(s):
+           print("{} is nonsense".format(s))
+        else:
+           print("{} is real".format(s))
 
-The value of `result` will be a Boolean, with the value `True` if the input string is probably meaningless and `False` if it is probably not.  (Note: the first time you import the module `nostril`, it will take extra time because Nostril loads a large data file during the import step.  If you only execute the short two-line example above, it will seem that Nostril takes far too long to evaluate a string.  This is misleading: one it's loaded, multiple calls to `is_nonsense()` are very fast.)
+The value returned by `nonsense` will be `True` if the input string is probably meaningless and `False` if it is probably not.  (Note: the first time you import the module `nostril`, it will take extra time because Nostril loads a large data file during the import step.  If you only execute the short two-line example above, it will seem that Nostril takes far too long to evaluate a string.  This is misleading: one it's loaded, multiple calls to `nonsense()` are very fast.)
 
-It is possible to tune some of the parameters used by the classifier.  The parameters are part of the mathematical function used internally by Nostril to compute a score for a given input string.  To get a pointer to a new classifier function (a closure) with different values of the tunable parameters, call `generate_nonsense_detector` like so:
-
-    from nostril import generate_nonsense_detector
-    is_nonsense = generate_nonsense_detector(...)
-
-where `...` are parameters that are explained in the help string.  The new function `is_nonsense()` obtained by calling the generator this way can be used exactly as the default version of `is_nonsense()`:
-
-    result = is_nonsense('yoursinglestringhere')
-
[…]
