### 2018-04-29 b3caf0d
Add new NEWS.md file
