+++ NEWS.md
@@ -1,0 +1,12 @@
+📰 Nostril News
+==============
+
+This file summarizes the main changes between different versions of this package.
+
+News for version 1.1.1
+----------------------
+
+* Fix [`requirements.txt`](requirements.txt) to use `>=` in version comparisons instead of exact comparisons
+* Fix and enhance the Nostril documentation in [docs/explanations/README.md](docs/explanations/README.md)
+* Fix a broken link in [`README.md`](README.md) to the explanations document 
+* Add a [NEWS.md](NEWS.md) file
