### 2018-04-29 dbe5427
Update README to mention changes in latest version
