+++ README.md
@@ -11,10 +11,14 @@ Nostril is the _Nonsense String Evaluator_: a Python module that infers whether 
 *Code repository*:   [https://github.com/casics/nostril](https://github.com/casics/nostril)<br>
 *License*:      Unless otherwise noted, this content is licensed under the [GPLv3](https://www.gnu.org/licenses/gpl-3.0.en.html) license.
 
+🏁 Recent news and activities
+------------------------------
+
+_April 2018_: Version 1.1.1 fixes an issue with `requirements.txt` that made it do exact version comparisons instead of doing minimum version comparisons.  It also updates the documentation in [docs/explanations](docs/explanations).  Other notable changes (which were in version 1.1.0) are a fix to an issue in `setup.py` that caused the installation process to fail to install dependencies automatically, with the consequence that users got errors about missing Python packages.  Additional improvements include updated installation instructions [below](#-installing-nostril), slight improvements to the [JOSS paper](docs/papers/joss/paper.pdf), a change to the command-line program to use the more conventional `-V` instead of `-v` for printing the version, and internal code refactoring.
+
 Table of contents
 -----------------
 
-* [Recent news and activities](#-recent-news-and-activities)
 * [Introduction](#-introduction)
 * [Installing Nostril](#-installing-nostril)
 * [Using Nostril](#-using-nostril)
@@ -24,11 +28,6 @@ Table of contents
 * [Getting help and support](#-getting-help-and-support)
 * [Contributing — info for developers](#-contributing--info-for-developers)
 * [Acknowledgments](#️-acknowledgments)
-
-🏁 Recent news and activities
-------------------------------
-
-_April 2018_: Version 1.1.1 fixes an issue with `requirements.txt` that made it do exact version comparisons instead of doing minimum version comparisons.  It also updates the documentation in [docs/explanations](docs/explanations).  Other notable changes (which were in the immediately-prior version 1.1.0) are a fix to an issue in `setup.py` that caused the installation process to fail to install dependencies automatically, with the consequence that users got errors about missing Python packages.  Additional improvements include updated installation instructions [below](#-installing-nostril), slight improvements to the [JOSS paper](docs/papers/joss/paper.pdf), a change to the command-line program to use the more conventional `-V` instead of `-v` for printing the version, and internal code refactoring.
 
 ☀ Introduction
 -----------------------------
