### 2018-04-29 30b0136
Fix "latest version" badge to have correct version number
