### 2018-05-06 #596 @arfon
:wave: @desilinguist - could you take another look when you get a chance?  I think we're close to being done here.
