### 2018-05-07 #596 @arfon
@jakevdp - are we good to accept here?
