### 2018-05-07 #596 @jakevdp
Yes, I think we're good to go!
