### 2018-05-11 #596 @mhucka
@arfon and @jakevdp and @desilinguist  – Thanks!

I used Caltech's DataCite-based DOI service this time around. The DOI is now [10.22002/D1.935](http://dx.doi.org/10.22002/D1.935) and should be operational.
