### 2018-05-11 62b15df
Add DOI for copy in data.caltech.edu
