### 2018-05-11 #596 @arfon
Thanks for your review @desilinguist and to @jakevdp for editing this one!

@mhucka - your paper is now accepted into JOSS and your DOI is http://joss.theoj.org/papers/10.21105/joss.00596
